requirements.txt
google-genai
python-dotenv
requests
streamlit
apify-client
supabase
# fake-useragent
# selenium
# playwright
# pillow

# Additional dependencies for Jupyter Notebook and IPython kernel
# ipykernel

# For Playwright, you may need to install the browsers separately
# playwright install